requirements.txt
sentence-transformers
transformers
scikit-learn
torch
requests
python-dotenv
